[agent] refactor(assembly): slim lib.rs entry + test extraction
